Move scope utilities to oauth
